Fix warnings in AddressesKit.  This is horrible code, which really needs replacing with the CoreObject version soon...
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@7919 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Frameworks/AddressesKit/Frameworks/AddressView/ADPersonPropertyView+Events.m
@@ -52,7 +52,7 @@
   
   if([c isKindOfClass: [ADPersonActionCell class]])
     {
-      ADActionType type; NSString *ident, *label; int index;
+      ADActionType type; NSString *ident, *label; NSInteger index;
 
       type = [c actionType];
